fix: enhance TurboQuant support and warp size handling in set-rows kernel
- Add support for D=512 in Turbo3, Turbo2, and Turbo4 vector kernels
- Ensure correct Turbo dequantization by forcing VEC kernel for Turbo K/V types
- Adjust __ballot_sync to use hardware warp size for sign packing in Turbo kernels
- Update macro definitions to use unsigned long long for __ballot_sync output

+++ ggml/src/ggml-cuda/vendors/hip.h
@@ -58,7 +58,7 @@
 #define __shfl_down_sync(...) __SHFL_DOWN_GET(__VA_ARGS__, __SHFL_DOWN_SYNC_4, __SHFL_DOWN_SYNC_3)(__VA_ARGS__)
 #define __all_sync(mask, var) __all(var)
 #define __any_sync(mask, var) __any(var)
-#define __ballot_sync(mask, var) ((uint32_t)__ballot(var))
+#define __ballot_sync(mask, var) ((unsigned long long)__ballot(var))
 #define cublasStrsmBatched hipblasStrsmBatched
 #define cublasCreate hipblasCreate
 #define cublasDestroy hipblasDestroy
